+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- add `qr` subcommand
+
 ## v0.2.0
 
 - log info/warning/error core events by default