+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - add `qr` subcommand
 - switch to zap logger
 - update configAction() to print a new line in the returned config value
+- panic if deltachat-rpc-server can't be started and provide hint to installation instructions
 
 ## v0.2.0